AwareAI/requirements (1).txt
asgiref==3.4.1
cffi==1.15.0
cryptography==35.0.0
Django==3.2.8
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
mysqlclient==2.0.3
pycparser==2.20
PyJWT==2.3.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
